+++ src/main/java/cn/wthee/pcrtool/ui/home/CharacterListFragment.kt
@@ -1,13 +1,10 @@
 package cn.wthee.pcrtool.ui.home
 
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.appcompat.app.AlertDialog
 import androidx.core.content.edit
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -21,9 +18,7 @@ import cn.wthee.pcrtool.MainActivity.Companion.sortType
 import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.adapter.CharacterListAdapter
 import cn.wthee.pcrtool.data.model.FilterCharacter
-import cn.wthee.pcrtool.database.DatabaseUpdater
 import cn.wthee.pcrtool.databinding.FragmentCharacterListBinding
-import cn.wthee.pcrtool.databinding.LayoutWarnDialogBinding
 import cn.wthee.pcrtool.enums.SortType
 import cn.wthee.pcrtool.utils.*
 import cn.wthee.pcrtool.utils.Constants.LOG_TAG
@@ -31,7 +26,6 @@ import com.google.android.material.transition.Hold
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
-import kotlin.system.exitProcess
 
 /**
  * 角色列表
@@ -44,7 +38,6 @@ class CharacterListFragment : Fragment() {
         var characterFilterParams = FilterCharacter(
             true, 0, 0, false, "全部"
         )
-        lateinit var handler: Handler
         lateinit var guilds: ArrayList<String>
         var r6Ids = listOf<Int>()
     }
@@ -90,66 +83,6 @@ class CharacterListFragment : Fragment() {
         setObserve()
         //获取角色
         viewModel.getCharacters(sortType, sortAsc, "")
-        //接收消息
-        handler = Handler(Looper.getMainLooper(), Handler.Callback {
-            when (it.what) {
-                //获取版本失败
-                0 -> {
-                    val layout = LayoutWarnDialogBinding.inflate(layoutInflater)
-                    //弹窗
-                    DialogUtil.create(
-                        requireContext(),
-                        layout,
-                        Constants.NOTICE_TITLE_ERROR,
-                        Constants.NOTICE_TOAST_TIMEOUT,
-                        Constants.BTN_OPERATE_FORCE_UPDATE_DB,
-                        Constants.BTN_NOT_UPDATE_DB,
-                        object : DialogListener {
-                            override fun onCancel(dialog: AlertDialog) {
-                                //强制更新数据库
-                                DatabaseUpdater.forceUpdate()
-                                ToastUtil.short(Constants.NOTICE_TOAST_TITLE_DB_DOWNLOAD)
-                                dialog.dismiss()
-                            }
-
-                            override fun onConfirm(dialog: AlertDialog) {
-                                dialog.dismiss()
-                            }
-                        }
-                    ).show()
-                }
-                //正常执行
-                1 -> {
-                    viewModel.reload.postValue(true)
-                }
-                //数据切换
-                2 -> {
-                    val layout = LayoutWarnDialogBinding.inflate(layoutInflater)
-                    //弹窗
-                    DialogUtil.create(
-                        requireContext(),
-                        layout,
-                        getString(R.string.change_success),
-                        getString(R.string.change_success_tip),
-                        getString(R.string.close_app),
-                        getString(R.string.close_app_too),
-                        object : DialogListener {
-                            override fun onCancel(dialog: AlertDialog) {
-                                requireActivity().finish()
-                                exitProcess(0)
-                            }
-
-                            override fun onConfirm(dialog: AlertDialog) {
-                                requireActivity().finish()
-                                exitProcess(0)
-                            }
-                        }
-                    ).show()
-                }
-            }
-
-            return@Callback true
-        })
         return binding.root
     }
