+++ src/main/res/layout/fragment_character_rank_compare.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
     android:paddingTop="@dimen/xm_padding"
     tools:context=".ui.detail.character.CharacterRankCompareFragment">
 
@@ -51,7 +50,6 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/m_margin"
-        android:animateLayoutChanges="true"
         android:orientation="horizontal"
         android:gravity="center"
         app:layout_constraintTop_toBottomOf="@id/rank_btns_1">
