+++ src/main/res/layout/fragment_character_skill.xml
@@ -10,7 +10,6 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:animateLayoutChanges="true"
         android:minHeight="@dimen/card_height">
 
         <include
