+++ src/main/res/layout/fragment_common_bottom_sheet.xml
@@ -4,6 +4,5 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
     android:background="@color/colorAlpha"
     tools:context=".ui.common.ContainerFragment" />
