+++ src/main/res/layout/fragment_tool_event.xml
@@ -5,7 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@drawable/background_color"
-    android:animateLayoutChanges="true"
     android:gravity="center"
     android:orientation="vertical">
