+++ src/main/res/layout/item_character_attr.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginTop="@dimen/m_margin"
-    android:animateLayoutChanges="true"
     android:orientation="horizontal">
 
     <com.google.android.material.textview.MaterialTextView
