+++ src/main/res/layout/item_character_drop.xml
@@ -7,7 +7,6 @@
     android:paddingStart="@dimen/xm_padding"
     android:paddingEnd="@dimen/xm_padding"
     android:paddingTop="@dimen/s_padding"
-    android:animateLayoutChanges="true"
     android:orientation="horizontal">
 
     <com.google.android.material.textview.MaterialTextView
