+++ src/main/res/layout/item_equipment_drop.xml
@@ -5,7 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="@dimen/s_padding"
-    android:animateLayoutChanges="true"
     android:orientation="vertical">
 
     <com.google.android.material.textview.MaterialTextView
