+++ src/main/res/layout/item_skill_action.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_margin="@dimen/xs_margin"
-    android:animateLayoutChanges="true"
     android:gravity="top"
     android:orientation="horizontal">
