+++ src/main/res/layout/layout_warn_dialog.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:animateLayoutChanges="true"
     android:minHeight="@dimen/card_height"
     android:orientation="vertical"
     android:paddingStart="@dimen/xm_padding"
