+++ src/main/res/layout/view_toolbar.xml
@@ -5,7 +5,6 @@
     android:id="@+id/view_toolbar"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:animateLayoutChanges="true"
     android:background="@color/colorAlpha"
     android:paddingStart="@dimen/xs_padding"
     android:paddingTop="@dimen/s_padding"
